requirements.txt
schedule~=1.2.0
PyMySQL~=1.1.0
python-dotenv~=1.0.0
pytz~=2023.3